chore: rename botonicapiservice with -
