+++ packages/botonic-cli/src/commands/login.ts
@@ -2,7 +2,7 @@ import { Command, flags } from '@oclif/command'
 import { prompt } from 'inquirer'
 import { resolve } from 'path'
 
-import { BotonicAPIService } from '../botonicapiservice'
+import { BotonicAPIService } from '../botonic-api-service'
 import { track } from '../utils'
 
 export default class Run extends Command {
